+++ lib/src/enum.dart
@@ -1,0 +1,1 @@
+enum TooltipPosition { top, bottom }